+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/utils/http/HttpResponseCallback.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.wazuh.contentmanager.util.http;
+package com.wazuh.contentmanager.utils.http;
 
 import org.apache.hc.client5.http.async.methods.SimpleHttpRequest;
 import org.apache.hc.client5.http.async.methods.SimpleHttpResponse;
