+++ plugins/content-manager/src/test/java/com/wazuh/contentmanager/utils/UnzipTests.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.wazuh.contentmanager.util;
+package com.wazuh.contentmanager.utils;
 
 import org.opensearch.common.settings.Settings;
 import org.opensearch.env.Environment;
